Introduce stdio `/dev/` devices to the file system (#79)
This PR introduces a new filesystem which is aware of exactly 3 files (`/dev/std{in,out,err}`). It is intended to be used with the layering system to provide these as devices in Linux-y environments. Non-Linux-y environments are unaffected by simply not including them into the layering tree.

Expected layering:
```mermaid
flowchart TD
    l["layered"] --> im["in_mem"]
    l --> l2["layered"]
    l2 --> d["dev::stdio"]
    l2 --> tr["tar_ro"]
```