De-parameterize Linux punchthroughs (#110)
This PR removes the generalized parametricity that we were keeping for
punchthrough for Linux-y LiteBox systems. The old reason for keeping
these was because we had not yet finalized `litebox_common_linux` or
what we actually needed from punchthroughs, or how the shim layer would
look like. At this point, we have a relatively good design for each of
them, thus we don't need the parametricity, and can instead put the
shared elements together in a nicer form, so that there is a clearer
indication of what needs to be punched through in order to support a
Linux-y north interface.