Improve setup for GitHub Copilot coding agent (#170)
By default, the coding agent doesn't seem to know that it should set up
(say) the TUN device, or install nextest, and it needs to spend time
figuring it out, or failing to do so. The
`.github/workflows/copilot-setup-steps.yml` sets these things up in
advance.

Additionally, it noticeably seems to forget to format code, so I've
added `.github/copilot-instructions.md` that tells it to do this for
each commit. I also gave it some more information, which should
hopefully send it down a nicer path that better matches workflow within
our repo.