Support kernel text patching (#189)
This PR adds initial support for patching kernel/module text. It
introduces the `mshv_vsm_patch_text` function that the VTL0 kernel calls
when it needs to patch kernel or module text regions (which are write
protected with SLAT). Currently, this function does not check whether a
patching request potentially violates the security (which will be
covered in the next PR). It only does sanity checks.