Provide more data for FS stat (#211)
This PR adds uid/gid/dev/ino/rdev/blksize to the `FileStatus` provided
by the file systems in LiteBox. This removes some of the specialization
that exists within the shim for stdio, and paves the way towards
switching the descriptor interfaces to being fully within LiteBox (as
part of the series begun by #180).

---

Note: as a drive-by change, the `FileSystem` trait had become unwieldy
with the lack of spaces between functions; this PR also just adds a
bunch of newlines there to improve readability.