ELF loader and Linux-userland runner for OP-TEE TAs (#246)
This PR adds initial implementation of the ELF loader and the
Linux-userland runner for OP-TEE TAs.

This TA ELF loader and `mmap*` functions are largely from
`litebox_shim_linux` except that
this loader does not assume a filesystem (everything is in memory) and
it invokes the `easy_relocate`
function (because there is no `ld`). It is able to reuse the `mmap*`
functions/syscalls in `litebox_shim_linux`
because OP-TEE OS does not directly expose memory management system
calls to TAs.
They are for LDELF (OP-TEE's ELF loader) and Pseudo TAs (i.e.,
kernel-mode TAs). To this end, it does not have to
implement separate `mmap*` functions/syscalls for TAs.

It also introduce a runner for OP-TEE TAs on the Linux userland
(`litebox_runner_optee_on_linux_userland`).
This runner is expected to ease development and debugging. Currently,
this runner does not support
end-to-end TA operation (i.e., open, invoke, close). It can load and
start a TA and intercept its syscalls, but
its interaction with TAs is incomplete (e.g., it lacks an event loop to
control the end-to-end operation) which
will be covered in the next PR.