Fix stack alignment for new threads in tests (#307)
A crash showed up in the unit test `test_thread_spawn` due to stack misalignment
when using SSE instructions.

The root cause is that the stack pointer at the entry of the new thread
is 16-byte aligned, but regular functions assume a 16-byte aligned stack
pointer right before the call instruction, which means the stack pointer
at the entry of a regular function is actually 8-byte aligned. We only
need to do this if we want to pass a Rust function to `sys_clone`.

This PR also fixes the features so that we can run each individual test.