Support for KMPP TA (#306)
This PR adds initial support for the KMPP TA. It partially implements
five OP-TEE syscalls (`GetProperty`, `GetPropertyNameToIndex`,
`OpenTaSession`, `CloseTaSession`, `InvokeTaCommand`) and
related data structures to provide some essential functionalities
that the KMPP TA assumes like
`PTA_SYSTEM_DERIVE_TA_UNIQUE_KEY` to generate a TA unique
key to protect the keys that the KMPP TA generates or imports
from a client app. We confirm that this PR works for a KMPP TA
built with `--no-asm` option and SymCrypt with a minor patch.