Use RefCell to protect TLS
Use a `RefCell` instead of manually implementing the same approach.
This moves the `borrowed` field out of the common TLS type, which
closes a soundness hole. Move `self_ptr` out for the same reason.

Also, on Windows, use linker-provided TLS instead of runtime calls.
This is more efficient and safer, and it works because Windows's TLS
machinery uses gs as a base, which is not used by user code. Once fs
is swapped across syscalls, we can use this approach for Linux and
FreeBSD as well.