Minor/style: allow clippy too-many-lines (#405)
It appears that we are not actively splitting functions at the specific
thresholds that `clippy::too_many_lines` uses and instead an `expect` on
it gets introduced as soon as it complains.

Having this pedantic lint around seems counter-productive, so this PR
just allows this pedantic lint through, and removes the attributes on
all the functions that were using it before.