Use jmp instead of call in syscall rewriter
On x86_64, store the return address in rcx instead of on the stack
when branching to the syscall callback. This is necessary because
the stack pointer may be invalid at the time of a syscall, such as
during thread teardown in binaries using musl libc.

For now, just fix all the syscall callbacks to handle this new
calling convention by pushing rcx to the stack on entry. Later changes
will update platforms to take care to never touch the guest stack
in the syscall path.