Support setitimer (#388)
Similar to #346, another signal related syscall
[`setitimer`](https://www.man7.org/linux/man-pages/man2/setitimer.2.html)
is added as punchthrough. We could potentially emulate it with a
background thread and `tgkill`, but I'd like to defer it until we figure
out how to handle signals. Adding this now so that we have a better
understanding how many signal related syscalls we should consider.