Make shim TLS type opaque to platform
Simplify the platform TLS trait to just operate on a single pointer,
moving the complexity of managing this pointer into litebox.

This simplifies the code the platform needs to implement. But more
importantly, this allows the shim to define its TLS type locally,
without making it visible outside the crate. This will be important
as we expand the complexity of the per-task state in the Linux shim.

Currently, only a single TLS value can be set on a thread at a time.
This is sufficient for the needs of the existing shims.