Centralize constructing initial context for new threads (#423)
The shim should fully construct the thread's initial context (including
stack pointer, instruction pointer, and return value) and not leave it
up to the individual platforms.

As part of this, fix `ThreadProvider` so that the thread arguments do
not leak internal shim details. This is straightforward now that the
thread ID is managed by the shim and not the platform.