Handle execve fully within the shim
Instead of calling out to the platform to load and run the next program,
handle the execve syscall mostly within the Linux shim (a callback
has been preserved to allow the Linux userland shim to update auxv and
envp.)

This is only possible now that #432 has landed.

Note that execve still does not work on Windows userland, apparently due
to some failure when unmapping memory.