Store context in platform memory, not on guest stack (#432)
Instead of pushing the guest context onto the guest stack, store it in
pre-allocated context memory, on the platform stack. This fixes issues
with musl on x86_64 where sometimes it issues syscalls without a valid
stack pointer.

This change also ensures that the guest context is accurately restored
in all cases that are possible. One exception is Windows, where we
return to the guest with rcx = rip; this matches the semantics of
sysret, so that's OK for returning from syscalls, but it doesn't work
for signal return. In that case, we use `NtContinue` to atomically set
the full context.

Another benefit is that the stack can now be unwound during a syscall,
so you can see the full stack if a syscall handler panics. This is
enabled by including unwind metadata in the asm routines, and by merging
the thread entry point with the syscall entry point, so that the
unwinder can see that they share a stack frame.