Add exception handler infrastructure
Add an interface for the platform to call the shim when the guest code
causes an unresolved hardware exception (page fault, invalid opcode,
divide by zero, etc.). Implement support for catching exceptions in the
Linux and Windows userland platforms.

For now, the Linux and OPTEE shims panic when an exception occurs. In
future changes, the Linux shim will inject the appropriate signal into
the current task in response to a signal. The correct OPTEE behavior
is less clear.