platform_lvbs: Fix for boot stalling if COM2 is not available
If COM2 is not enabled, on a print, a cpu in VTL1 loops infinitely
trying to access COM2. Fix this by a) returning if COM2 is not available
instead of looping b) adding a timeout to the loop so that a CPU exits out
of it in case of hardware failure.