Clear out manually-handled pipes from the shim (#451)
This PR switches out the manually handled pipes in the shim and instead
uses LiteBox raw FDs for them.

It additionally switches away the old internal details from
`litebox::pipes` to be private, so that we can use the more
consistent-with-other-subsystems interface introduced by #446, which
will then allow future cleanups to happen inside `litebox::pipes`.

---

This is yet another PR in the series of PRs towards the better FD design
(connected to https://github.com/microsoft/litebox/issues/31).