Add interrupt support to userland platforms
Add a mechanism to interrupt a thread that's running in the guest so
that the shim can inject a signal or terminate the thread. On the
Linux platform, use a signal to interrupt the target thread. On Windows,
use `GetThreadContext`/`SetThreadContext`.

This is not actually hooked up to any shims yet in this PR, but I have
tested this code somewhat via some private changes.