Improve runner test output
In the runner tests that launch the runner as a binary, instead of
capturing stdout/stderr, inherit them. This ensures that stderr (and
stdout, for most tests) is visible in CI even if the test crashes or
times out.

Also, skip the cargo invocation since the binary is guaranteed by
cargo to be built when the tests run. Use $CARGO to launch cargo for
other binaries.

Refactor the Linux runner test executor to use a builder pattern.

Finally, clean up duplicated code in the optee runner tests.