Fix fixed address memory map race conditions
When the page manager is asked to perform an allocation at a fixed
address, it first unmaps any old allocations before putting the new
one in place. If the platform allocates memory on another thread between
these two steps, then the desired address may be reused, causing memory
corruption.

Fix this by telling the platform whether to expect an existing
mapping to be present when allocating pages, and rely on the platform
to overwrite those ranges atomically--don't unmap (and thereby
release the VA ranges of) the existing mappings first.