Add infrastructure for interruptible waits
To support interrupted and terminated threads that are waiting on
guest- or externally-triggered events, add infrastucture for performing
interruptible waits. Each thread stores a `WaitState`, which can be
used to generate `WaitContext` objects to wait on a condition, plus
`Waker` objects that can be used to cause a thread to unblock to
reevaluate the condition.

Also add a `ThreadHandle` object to allow interrupting a thread whether
it is waiting or running in the guest.

Update all IO-related code to use this new wait infrastructure.

Futex has yet to be updated since it must be rewritten to work with
a wait-on-condition style of waiting, and that should be reviewed
separately.