Fix soundness issues with Mutex and RwLock
`Mutex<T>` can only be soundly `Send`/`Sync` if `T` is `Send`. Similar
for `RwLock<T>` (although in that case, `Sync` requires `T: Sync`).
These bounds were missing, which created a soundness hole.

Luckily, it doesn't appear that any undefined behavior crept in. This
change just adds the appropriate bounds, which had to be propagated
to a few other traits as well.