Manually load ELF binaries
Parse and load ELF binaries manually instead of using the `elf_loader`
crate. This is done for a few reasons:

* `elf_loader` is not safe for our use--it assumes it can safely access
  mapped memory directly via Rust memory access primitives--this is
  not always true for LiteBox since that memory is user-mode memory.
  Even when we know there are no other guest threads, so that concurrent
  mutation should be impossible, we still need to handle that memory
  differently from Rust memory.

* `elf_loader` creates a dependency on shim TLS to get the task object,
  since the various mmap-related traits do not take `&self`. This
  prevents some future refactoring to eliminate the use of TLS within
  the shims. Fixing this upstream seems like a big change, one
  that would break semver compatibility.

* `elf_loader` seems to combine parsing and mapping operations into
  one step, preventing us from returning errors from `execve` when
  the target binary is not a valid ELF executable.

* `elf_loader` has an awful lot of code marked unsafe with unclear
  safety contracts.

* Loading the binary is a pretty important part of the Linux and OPTEE
  shims--we want to have a good understanding and control over exactly
  what loader ABI we are actually implementing.

As part of this, implement relocation support for OPTEE, since before
we were relying on `elf_loader` for this--this is not necessary for
the Linux shim. I am wondering if in the future we can instead inject
a loader binary into OPTEE user mode, similar to the interpreter used
by Linux binaries; then we can run the relocator and whatever else we
want (e.g., TLS allocation) there.

he commit message for your changes. Lines starting