Update the OPTEE shim to use `CrngProvider`
The OPTEE shim should use the new provider to generate random bytes--I
missed this in my previous PR. As part of this, fix the shim's handling
of large RNG requests to match OPTEE OS (failing above 4096 bytes).

Also, add the fake RNG implementation to the LVBS kernel, to be replaced
later.