Clean up `ContinueOperation` (#502)
Some clean up in preparation for more task/signal work:

* Define one `ContinueOperation` for all shims, with just
`ContinueGuest` and `ExitThread`.

* Remove exit status from `ExitThread`--this, plus the distinction
between thread and process exit, will be communicated separately in
future PRs, on a per-shim basis.

* Unify Linux shim entry/exit code.

* Record the `is_exiting` state of a Linux thread in `Task` and use it
to determine whether to return `ExitThread` from a shim entrypoint. This
is a more natural fit for Linux signal/termination handling.