More thread startup cleanup
Make Linux shim entrypoint objects to be per-thread and not global, and
update the platforms to handle this. This eliminates the global shim
registration, which in theory would allow multiple shims to be used
at the same time (on different threads, of course).

For now, there is still a bit of a TLS dance, which should go away
once we stop using TLS in the Linux shim.

The OPTEE shim keeps its current model for now.