Fix rewriter output in some edge cases
The rewriter calls into `object` to recomputes PT_LOAD segment sizes
from the section headers before writing the new binary; in some cases,
this truncates out data at the beginning of the segment that is not
included in any sections. Specifically, if the trampoline section gets
put in the first segment, this sometimes strips out the program headers,
preventing the binary from being loaded by Linux.

Fix this by skipping the recomputation. The segment length is already
extended by `object` when the trampoline section is added, so this
recomputation is not necessary.