Make `RawMutex`, `Mutex`, and `RwLock` const constructible
Currently, constructing a `RawMutex` required a `Platform` object.
Similarly, constructing a `Mutex` or `RwLock` requires a
`Synchronization` object. This is inconvenient for some use cases:

* It's not straightforward to initialize a mutex of any kind in a const
  context. Doing so would be helpful for avoiding runtime initialization
  overhead for some globals, and for defining a LiteBox version of
  `OnceLock` for lazy initialization of other globals. This would allow
  us to eliminate some uses of `spin::Once` and `once_cell::race`.
* It's hard to use different mutex types in the same project. This would
  be useful, for example, to allow using `LoanList<SpinRawMutex>` or
  `FutexManager<SpinRawMutex>` in a platform implementation of
  a wait-based `RawMutex`.
* It's hard to implement `Clone` for types containing mutexes--you have
  to preserve the `&'static Platform` pointer to make this possible,
  wasting space, or you have to implement `clone_with_platform`-style
  functions. The latter can't be used with `Arc::make_mut()`, etc.

These problems go away if the mutex types can be constructed out of thin
air and in a const context.

Toward that end, this change updates the `RawMutex` trait to contain an
`const INIT: Self` member that contains the const-initialized state of
a `RawMutex`. It updates `Mutex` and `RwLock` to have
`pub const fn new()` methods that take advantage of this.

To facilitate this, this changes updates the lock tracker to be a
singleton, initialized at runtime with the platform to use for
timekeeping and log output. This singleton is used for all
`Mutex`/`RwLock` instances, regardless of whether they use the
platform's `RawMutex`. This is slightly weird in the presence of
multiple platform instances, but I don't think that's a real problem in
practice.

This approach puts a small burden on platforms to ensure their
`RawMutex` types are const-constructible without having access to
the underlying `Platform` instance. In practice, I have not found
any implementations where this is a problem, nor do I anticipate any
such implementations. Intra-process synchronization primitives are a
place where a capabilities-based approach doesn't seem to add much
value.

So, although there may be other ways to solve the problems enumerated
above, this seems like the right tradeoff.

To keep this PR scoped, I have not yet updated any callers to leverage
these changes. A future change can eliminate the use of `litebox.sync()`
across the codebase, eliminating a bunch of references to the `LiteBox`
object.