Add lifetime to punchthrough tokens
Add a lifetime to punchthrough tokens, propagating through the
punchthrough provider trait. This is useful for passing references
through a punchthrough token, allowing us to avoid using a mut pointer
and an extra unsafe block on x86-32.

Also, take the opportunity to change `GetFsBase` to return its value
instead of writing it to a pointer/reference.

Finally, add an uninhabited variant to `PunchthroughSyscall` so that
generic parameters that are only used on some architectures don't
cause compile errors.