Improved UX for litebox_rtld_audit.so (#540)
This PR makes two changes:
- uses an explicit Makefile, rather than just relying on the `build.rs`,
making it easier to test builds (and if needed, have a way to build to a
consistent path)
- switches to auto-including the `litebox_rtld_audit.so` file in the
in-mem FS rather than expecting the users to fill it into the tar file.
This improves the UX for users of the runner, and also prevents the
implicit dependency of path (in the future, we can move the rtld audit
file around). This _especially_ means that static binary users will now
have a much easier time running things (they may not even need to put
any tar file in).