Add unix datagram sockets (#548)
Add support for unix datagram socket and some unit tests similar to unix
stream sockets.

Also fix an issue on double binding on the same address. Previously, I
had a C test that showed it was okay to double bind on the same address.
Turns out it was because of `unlink` that I overlooked. Now use
`OFlags::EXCL` to ensure exclusive creation of sock file to address the
issue.